ajustes en la base de datos para mejores calculos

+++ back-end/apps/finanzas/stock/models.py
@@ -13,7 +13,7 @@ class Stock(models.Model):
     fk_id_produccion = models.ForeignKey(Produccion, on_delete=models.SET_NULL, null=True, blank=True)
     fk_id_venta = models.ForeignKey(Venta, on_delete=models.SET_NULL, null=True, blank=True)
     
-    cantidad = models.PositiveIntegerField()
+    cantidad = models.DecimalField(max_digits=20, decimal_places=3)
     fecha = models.DateTimeField(auto_now_add=True)
     movimiento = models.CharField(max_length=10, choices=MOVIMIENTO_CHOICES)
