+++ back-end/apps/finanzas/venta/models.py
@@ -4,16 +4,15 @@ from apps.inventario.unidadMedida.models import UnidadMedida
 # Create your models here.
 
 class Venta(models.Model):
-    precio_unidad = models.DecimalField(max_digits=10, decimal_places=2)  
-    cantidad = models.IntegerField()
-    fecha = models.DateField()  
-    fk_id_produccion = models.ForeignKey(Produccion, on_delete=models.SET_NULL, null=True) 
-    fk_unidad_medida = models.ForeignKey(UnidadMedida, on_delete=models.SET_NULL, null= True)
+    precio_unidad = models.DecimalField(max_digits=10, decimal_places=2)
+    cantidad = models.DecimalField(max_digits=20, decimal_places=0)  # Puede tener decimales
+    fecha = models.DateField()
+    fk_id_produccion = models.ForeignKey(Produccion, on_delete=models.SET_NULL, null=True)
+    fk_unidad_medida = models.ForeignKey(UnidadMedida, on_delete=models.SET_NULL, null=True)
 
-    # Cantidad convertida a unidad base (g, ml o unidad)
     cantidad_en_base = models.DecimalField(
         max_digits=20,
-        decimal_places=10,
+        decimal_places=3,
         null=True,
         blank=True,
         help_text="Cantidad vendida convertida a unidad base."
